requirements.txt
tensorflow==2.16.1
uvicorn==0.23.2
typing-inspect==0.9.0
typing_extensions==4.7.1

# ensorflow
scikit-learn
xgboost
mlflow
pandas
numpy
seaborn
matplotlib
scipy

Flask
gunicorn
Flask-Cors
werkzeug
requests
jinja2

PyAudio
speechRecognition
pyttsx3

dvc
python-dotenv
pyYAML
tqdm
joblib
ensure
uvicorn
dill
mlflow
notebook
python-box
types-PyYAML
gdown

-e .